+++ app/components/bulma/subtitle_component.rb
@@ -2,13 +2,14 @@
 
 module Bulma
   class SubtitleComponent < Component
-    def initialize(text = nil, size: nil)
+    def initialize(text = nil, as: :p, size: nil)
+      @as = as
       @text = text
       @size = size
     end
 
     def call
-      content_tag :p, @text || content, class: class_names("subtitle", "is-#{@size}" => @size.present?)
+      content_tag @as, @text || content, class: class_names("subtitle", "is-#{@size}" => @size.present?)
     end
   end
 end